// src/gen-ts/src/models/EquipmentIncident.ts
export interface EquipmentIncident {
    /**
     * 
     * @type {string}
     * @memberof EquipmentIncident
     */
    comment?: string;
    /**
     * 
     * @type {Date}
     * @memberof EquipmentIncident
     */
    created_at?: Date;
    /**
     * 
     * @type {Date}
     * @memberof EquipmentIncident
     */
    updated_at?: Date;
    /**
     * 
     * @type {AuditUser}
     * @memberof EquipmentIncident
     */
    created_by?: AuditUser;
    /**
     * 
     * @type {AuditUser}
     * @memberof EquipmentIncident
     */
    updated_by?: AuditUser;
    /**
     * 
     * @type {string}
     * @memberof EquipmentIncident
     */
    id?: string;
    /**
     * 
     * @type {IncidentType}
     * @memberof EquipmentIncident
     */
    incident_type?: IncidentType;
    /**
     * 
     * @type {string}
     * @memberof EquipmentIncident
     */
    equipment_id?: string;
    /**
     * 
     * @type {string}
     * @memberof EquipmentIncident
     */
    user_id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EquipmentIncident
     */
    travel_id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EquipmentIncident
     */
    location?: string | null;
}