// src/gen-ts/src/models/CrupdateEquipmentIncident.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni. 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { IncidentType } from './IncidentType';
import {
    IncidentTypeFromJSON,
    IncidentTypeFromJSONTyped,
    IncidentTypeToJSON,
    IncidentTypeToJSONTyped,
} from './IncidentType';

/**
 * 
 * @export
 * @interface CrupdateEquipmentIncident
 */
export interface CrupdateEquipmentIncident {
    /**
     * 
     * @type {string}
     * @memberof CrupdateEquipmentIncident
     */
    comment?: string;
    /**
     * 
     * @type {string}
     * @memberof CrupdateEquipmentIncident
     */
    id: string;
    /**
     * 
     * @type {IncidentType}
     * @memberof CrupdateEquipmentIncident
     */
    incident_type: IncidentType;
    /**
     * 
     * @type {string}
     * @memberof CrupdateEquipmentIncident
     */
    equipment_id: string;
    /**
     * 
     * @type {string}
     * @memberof CrupdateEquipmentIncident
     */
    user_id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CrupdateEquipmentIncident
     */
    travel_id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CrupdateEquipmentIncident
     */
    location?: string | null;
}



/**
 * Check if a given object implements the CrupdateEquipmentIncident interface.
 */
export function instanceOfCrupdateEquipmentIncident(value: object): value is CrupdateEquipmentIncident {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('incident_type' in value) || value['incident_type'] === undefined) return false;
    if (!('equipment_id' in value) || value['equipment_id'] === undefined) return false;
    return true;
}

export function CrupdateEquipmentIncidentFromJSON(json: any): CrupdateEquipmentIncident {
    return CrupdateEquipmentIncidentFromJSONTyped(json, false);
}

export function CrupdateEquipmentIncidentFromJSONTyped(json: any, ignoreDiscriminator: boolean): CrupdateEquipmentIncident {
    if (json == null) {
        return json;
    }
    return {
        
        'comment': json['comment'] == null ? undefined : json['comment'],
        'id': json['id'],
        'incident_type': IncidentTypeFromJSON(json['incident_type']),
        'equipment_id': json['equipment_id'],
        'user_id': json['user_id'] == null ? undefined : json['user_id'],
        'travel_id': json['travel_id'] == null ? undefined : json['travel_id'],
        'location': json['location'] == null ? undefined : json['location'],
    };
}

export function CrupdateEquipmentIncidentToJSON(json: any): CrupdateEquipmentIncident {
    return CrupdateEquipmentIncidentToJSONTyped(json, false);
}

export function CrupdateEquipmentIncidentToJSONTyped(value?: CrupdateEquipmentIncident | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'comment': value['comment'],
        'id': value['id'],
        'incident_type': IncidentTypeToJSON(value['incident_type']),
        'equipment_id': value['equipment_id'],
        'user_id': value['user_id'],
        'travel_id': value['travel_id'],
        'location': value['location'],
    };
}