// src/gen-ts/src/models/IncidentType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni. 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const IncidentType = {
    DAMAGED: 'DAMAGED',
    LOST: 'LOST'
} as const;
export type IncidentType = typeof IncidentType[keyof typeof IncidentType];


export function instanceOfIncidentType(value: any): boolean {
    for (const key in IncidentType) {
        if (Object.prototype.hasOwnProperty.call(IncidentType, key)) {
            if (IncidentType[key as keyof typeof IncidentType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function IncidentTypeFromJSON(json: any): IncidentType {
    return IncidentTypeFromJSONTyped(json, false);
}

export function IncidentTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): IncidentType {
    return json as IncidentType;
}

export function IncidentTypeToJSON(value?: IncidentType | null): any {
    return value as any;
}

export function IncidentTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): IncidentType {
    return value as IncidentType;
}